regen test data, make seed factor 64 bit

+++ src/py21cmfast/src/Stochasticity.c
@@ -1086,7 +1086,7 @@ int stochastic_halofield(unsigned long long int seed, float redshift_desc, float
 
     // set up the rng
     gsl_rng *rng_stoc[simulation_options_global->N_THREADS];
-    int seed_fac = (int)(redshift * 1000);
+    unsigned long long int seed_fac = (unsigned long long int)(redshift * 1000);
     seed_rng_threads_fast(rng_stoc, seed + seed_fac);
 
     struct HaloSamplingConstants hs_constants;
